09. Combine Categories & Posts Observables using CombineLatest & ForkJoin Operators
