+++ app/src/app/services/DeclarativeCategory.service.ts
@@ -1,0 +1,25 @@
+import { HttpClient } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+import { map } from 'rxjs';
+import { ICategory } from 'src/app/models/ICategory';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class DeclarativeCategoryService {
+  categories$ = this.http
+    .get<{ [id: string]: ICategory }>(
+      `https://rxjs-posts-default-rtdb.firebaseio.com/categories.json`
+    )
+    .pipe(
+      map((categories) => {
+        let categoriesData: ICategory[] = [];
+        for (let id in categories) {
+          categoriesData.push({ ...categories[id], id });
+        }
+        return categoriesData;
+      })
+    );
+
+  constructor(private http: HttpClient) {}
+} // The End of Class;
