--- app/src/app/services/DeclaritivePost.service.ts
@@ -1,25 +1,0 @@
-import { HttpClient } from '@angular/common/http';
-import { Injectable } from '@angular/core';
-import { map } from 'rxjs';
-import { IPost } from 'src/app/models/IPost';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class DeclarativePostService {
-  posts$ = this.http
-    .get<{ [id: string]: IPost }>(
-      `https://rxjs-posts-default-rtdb.firebaseio.com/posts.json`
-    )
-    .pipe(
-      map((posts) => {
-        let postsData: IPost[] = [];
-        for (let id in posts) {
-          postsData.push({ ...posts[id], id });
-        }
-        return postsData;
-      })
-    );
-
-  constructor(private http: HttpClient) {}
-} // The End of Class;
